Refactor to move claimTxId at utxo level. Misc fixes
