+++ .prettierrc.js
@@ -1,0 +1,4 @@
+module.exports = {
+  ...require('@ionic/prettier-config'),
+  printWidth: 80,
+};
