+++ src/utils/helpers.ts
@@ -20,7 +20,12 @@ import { lockUtxo } from '../redux/actions/walletActions';
 import { network } from '../redux/config';
 
 import type { AssetConfig, LbtcDenomination } from './constants';
-import { defaultPrecision, getColor, getMainAsset } from './constants';
+import {
+  defaultPrecision,
+  getColor,
+  getMainAsset,
+  LBTC_ASSET,
+} from './constants';
 import type { TxDisplayInterface } from './types';
 
 export const createColorFromHash = (id: string): string => {
@@ -252,9 +257,7 @@ export function customCoinSelector(dispatch?: Dispatch): CoinSelector {
 }
 
 export function getAssetHashLBTC(): string {
-  if (network.chain === 'regtest')
-    return '5ac9f65c0efcc4775e0baec4ec03abdde22473cd3cf33c0419ca290e0751b225';
-  return '6f0279e9ed041c3d710a9f57d0c02928416460c4b722ae3457a11eec381c526d';
+  return LBTC_ASSET.assetHash;
 }
 
 export function isLbtc(asset: string): boolean {
