+++ test/reducer/btc.reducer.spec.ts
@@ -91,21 +91,44 @@ describe('Btc reducer', () => {
         state,
         upsertPegins({
           '00146a010acf97d317e0c22cc622c789dbee7f40876c': {
-            claimTxId:
-              '334a48f635942e3a35c284b9787144826bed4692a2423005513490540dbb7dff',
             depositAddress: {
               claimScript: '00146a010acf97d317e0c22cc622c789dbee7f40876c',
               address: 'bcrt1qkcdq27w3zy5g8zekwreuzp4n9vd2jryjnql7dq',
               derivationPath: "m/84'/0'/0'/0/1",
             },
+            depositUtxos: {
+              '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff:1':
+                {
+                  txid: '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff',
+                  vout: 1,
+                  status: {
+                    confirmed: false,
+                  },
+                  value: 200000000,
+                },
+            },
           },
           '0014d863bb0f322219f3b0647a21d17ed44bf8fa9ae8': {
-            claimTxId:
-              'f5cbb4927c47bc34e41810948a10ca0304a7f293c6c1053058a6d1b3d0329bbd',
             depositAddress: {
               claimScript: '0014d863bb0f322219f3b0647a21d17ed44bf8fa9ae8',
               address: 'bcrt1qjg2rhft36x2zfskmndvmelmdzf4dxfq5ud0953',
               derivationPath: "m/84'/0'/0'/0/0",
+            },
+            depositUtxos: {
+              '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff:1':
+                {
+                  txid: '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff',
+                  vout: 1,
+                  status: {
+                    confirmed: true,
+                    block_height: 102,
+                    block_hash: '',
+                    block_time: 121212,
+                  },
+                  value: 200000000,
+                  claimTxId:
+                    'f5cbb4927c47bc34e41810948a10ca0304a7f293c6c1053058a6d1b3d0329bbd',
+                },
             },
           },
         }),
@@ -121,21 +144,44 @@ describe('Btc reducer', () => {
           },
         },
         '00146a010acf97d317e0c22cc622c789dbee7f40876c': {
-          claimTxId:
-            '334a48f635942e3a35c284b9787144826bed4692a2423005513490540dbb7dff',
           depositAddress: {
             claimScript: '00146a010acf97d317e0c22cc622c789dbee7f40876c',
             address: 'bcrt1qkcdq27w3zy5g8zekwreuzp4n9vd2jryjnql7dq',
             derivationPath: "m/84'/0'/0'/0/1",
           },
+          depositUtxos: {
+            '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff:1':
+              {
+                txid: '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff',
+                vout: 1,
+                status: {
+                  confirmed: false,
+                },
+                value: 200000000,
+              },
+          },
         },
         '0014d863bb0f322219f3b0647a21d17ed44bf8fa9ae8': {
-          claimTxId:
-            'f5cbb4927c47bc34e41810948a10ca0304a7f293c6c1053058a6d1b3d0329bbd',
           depositAddress: {
             claimScript: '0014d863bb0f322219f3b0647a21d17ed44bf8fa9ae8',
             address: 'bcrt1qjg2rhft36x2zfskmndvmelmdzf4dxfq5ud0953',
             derivationPath: "m/84'/0'/0'/0/0",
+          },
+          depositUtxos: {
+            '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff:1':
+              {
+                txid: '838228b25bf88186b7e95912afce225a0f05227cacc535e48b3d1030f93c0bff',
+                vout: 1,
+                status: {
+                  confirmed: true,
+                  block_height: 102,
+                  block_hash: '',
+                  block_time: 121212,
+                },
+                value: 200000000,
+                claimTxId:
+                  'f5cbb4927c47bc34e41810948a10ca0304a7f293c6c1053058a6d1b3d0329bbd',
+              },
           },
         },
       },
